:ambulance: :: count Type 을 Long -> Int로 변경
:

+++ src/main/kotlin/com/example/gungjeonjegwa/domain/basket/data/dto/BasketDto.kt
@@ -10,7 +10,7 @@ class BasketDto(
     val extramoney: Long?,
     val unit: String?,
     val isSoldOut: Boolean,
-    val count: Long,
+    val count: Int,
     val remaincount: Long,
 ) {
 }
